// aforte.next/src/modules/charity/interfaces/charity.d.ts
export type CharityT = {
  id: string;
  image: string;
  title: string;
  date?: string;
  time?: string;
  view?: string;
};

export type CharityDetailsT = {
  id: string;
  title: string;
  label: string;
  image: string;
  description: string;
  date?: string;
  time?: string;
  view?: string;
};